add newton approximation option for polar decomp
see https://doi.org/10.1137%2F070699895

the rescaling "zeta" is chosen to get the steps of the approximation
to be as orthogonal as possible.
that is, we prefer that approximation error make us inaccurate in diffusion direction
alignment rather than having it make us inaccurate in preserving eigenvalues
(i.e. diffusion amounts).